scripts/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
bandit==1.7.5
bleach==6.0.0
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.5
click-log==0.4.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.0
cryptography==41.0.2
ctranslate2==3.16.1
cymem==2.0.7
dataclasses-json==0.5.9
deprecation==2.1.0
doctran==0.0.9
docutils==0.20.1
dotty-dict==1.3.1
et-xmlfile==1.1.0
faster-whisper==0.6.0
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
invoke==1.7.3
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.1
jq==1.4.1
keyring==24.2.0
langchain==0.0.233
langcodes==3.3.0
langsmith==0.0.5
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pathy==0.10.2
pbr==5.11.1
pdf2image==1.16.3
pdfminer.six==20221105
phonenumbers==8.13.16
Pillow==10.0.0
pkginfo==1.9.6
postgrest==0.10.6
preshed==3.0.8
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
protobuf==4.23.4
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.11
Pygments==2.15.1
pypandoc==1.11
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-gitlab==3.15.0
python-magic==0.4.27
python-pptx==0.6.21
python-semantic-release==7.33.2
pytz==2023.3
PyYAML==6.0
readme-renderer==40.0
realtime==1.0.0
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.4.2
semver==2.13.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.19
srsly==2.4.6
stevedore==5.1.0
storage3==0.5.2
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
thinc==8.1.10
tiktoken==0.3.3
tldextract==3.4.4
tokenizers==0.13.3
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.8.1
urllib3==2.0.3
wasabi==1.1.2
webencodings==0.5.1
websockets==10.4
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.2
